### 2019-09-22 #1755 @whedon
[PRE REVIEW]: Hasasia: A Python package for Pulsar Timing Array Sensitivity Curves
**Submitting author:** @Hazboun6 (<a href="http://orcid.org/0000-0003-2742-3321">Jeffrey Hazboun</a>)
**Repository:** <a href="https://github.com/Hazboun6/hasasia" target ="_blank">https://github.com/Hazboun6/hasasia</a>
**Version:** v1.0.0
