### 2019-09-22 #1755 @arfon
:wave: @Hazboun6 - is this submission accompanying a AAS paper by any chance?
