### 2019-09-23 #1755 @Hazboun6
Hi @arfon, This is not associated with a AAS paper. There was a PRD
manuscript submitted (https://arxiv.org/abs/1907.04341) about the physics
that this code is based on. This was included as the citation which doesn't
seem to be working in the repo.

On Sun, Sep 22, 2019 at 4:38 PM Arfon Smith <notifications@github.com>
wrote:

> 👋 @Hazboun6 <https://github.com/Hazboun6> - is this submission
> accompanying a AAS paper by any chance?
>
