### 2019-09-23 #1755 @arfon
OK thanks.

The citation isn't being picked up because the citation keys don't match (you need to remove the `a` from the end of this key https://github.com/Hazboun6/hasasia/blob/master/paper/paper.bib#L25).

Could you please suggest a few potential reviewers for this submission? These can be people you know of in the community or from [this list](https://bit.ly/joss-reviewers).
