+++ paper/paper.bib
@@ -22,7 +22,7 @@
      adsurl = {http://adsabs.harvard.edu/abs/2013A%26A...558A..33A}
  }
 
- @article{hazboun:2019a,
+ @article{hazboun:2019,
          author         = "Hazboun, Jeffrey S. and Romano, Joseph D. and Smith, Tristan L.",
          title          = "{Realistic sensitivity curves for pulsar timing arrays}",
          year           = "2019",
