### 2019-09-23 #1755 @Hazboun6
Thanks, I've just fixed the typo. Just out of curiosity, is there a way to
check typesetting of the paper before submission?

From the list of reviewers here is a list of suggestions in roughly
prioritized order by the subject matter of the code's application.
@cmbiwer, @duncanmmacleod, @nespinoza, @FaustinCarter, @shreyasbapat

On Mon, Sep 23, 2019 at 4:35 AM Arfon Smith <notifications@github.com>
wrote:

> OK thanks.
>
> The citation isn't being picked up because the citation keys don't match
> (you need to remove the a from the end of this key
> https://github.com/Hazboun6/hasasia/blob/master/paper/paper.bib#L25).
>
> Could you please suggest a few potential reviewers for this submission?
> These can be people you know of in the community or from this list
> <https://bit.ly/joss-reviewers>.
>
