### 2019-09-23 27fed81
Added functions to load NG sensitivity curves.
