+++ hasasia/sensitivity.py
@@ -4,10 +4,14 @@ from __future__ import print_function
 import numpy as np
 import itertools as it
 import scipy.stats as sps
-import pickle
+import os, pickle
 from astropy import units as u
 
+import hasasia
 from .utils import create_design_matrix
+
+current_path = os.path.abspath(hasasia.__path__[0])
+sc_dir = os.path.join(current_path,'sensitivity_curves/')
 
 __all__ =['GWBSensitivityCurve',
           'DeterSensitivityCurve',
@@ -861,3 +865,26 @@ def make_quant(param, default_unit):
         quantity = param * default_unit
 
     return quantity
+
+################## Pre-Made Sensitivity Curves#############
+def nanograv_11yr_deter():
+    '''
+    Returns a `DeterSensitivityCurve` object built using with the NANOGrav
+    11-year data set.
+    '''
+    path = sc_dir + 'nanograv_11yr_deter.sc'
+    with open(path, "rb") as fin:
+        sc = pickle.load(fin)
+        sc.filepath = path
+    return sc
+
+def nanograv_11yr_stoch():
+    '''
+    Returns a `GWBSensitivityCurve` object built using with the NANOGrav 11-year
+    data set.
+    '''
+    path = sc_dir + 'nanograv_11yr_stoch.sc'
+    with open(path, "rb") as fin:
+        sc = pickle.load(fin)
+        sc.filepath = path
+    return sc
