### 2019-09-24 #1755 @arfon
> Thanks, I've just fixed the typo. Just out of curiosity, is there a way to
> check typesetting of the paper before submission?

I'm afraid there isn't currently although we're working on some additional tooling to make it easier for submitting authors.

@cmbiwer, @duncanmmacleod - would you both be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
