### 2019-09-29 73d693b
Added curves to __all__
