### 2019-10-01 #1755 @matteobachetti
Hi @arfon, sure.
It's the first time I do that for JOSS, could you point me to the guidelines?
