### 2019-10-01 #1755 @arfon
> It's the first time I do that for JOSS, could you point me to the guidelines?

Great, thanks so much! I'll go ahead and set up the review now and link to the guidelines there.
