### 2019-10-01 #1775 @matteobachetti
Overall, a very useful software package, well documented. I'm not done with the review, but I can start putting down some thoughts as I go through the checklist.

- Re. the statement of need: I ticked it because I think it's pretty clear what the software does and who can benefit, even if the audience is not *explicitly* mentioned.

- Tests are run on Travis. I'm verifying that they cover a significant amount of code.

- There are no community guidelines. In general, I think it's a good idea to have one.

- In the proof, there is one broken reference ( _on the formalism presented in (???)_ ). I would suggest to add two more references:

  1. a review on pulsar timing arrays in the introduction

  2. the newest Astropy paper (https://www.astropy.org/acknowledging.html)
