### 2019-10-01 #1775 @matteobachetti
Coverage is not bad. There are quite a few seemingly important pieces of the code left untested in `sensitivity.py`, including many of the methods in the `Spectrum` and `Sensitivity` classes. Even though I wouldn't block the publication based on that, I recommend to cover at least the ones you think users will rely the most on :).
