### 2019-10-01 #1775 @matteobachetti
Bottom line of my review: this is good code and deserves publication on JOSS. Before publication, I suggest to fix and improve the references as reported above (fix the broken one, add Astropy 2018 and a review on PTAs), and add some community/code of conduct guidelines to the repository.

As a further suggestion, it would be good to improve the test coverage, covering at least all important classes /methods that people will rely the most on.
