### 2019-10-01 #1775 @Hazboun6
@matteobachetti thanks for the quick/thorough review! I really like the real time refereeing. I assume it's kosher for me to do some real time fixes?

One question. For community guidelines, the docs have [this](https://hasasia.readthedocs.io/en/latest/contributing.html) standard template blurb about contributing. Is the issue that it's buried in the docs, i.e. that I should make it more front and center, or that I should expand on what's in there?
