### 2019-10-01 #1775 @matteobachetti
Sorry for the confusion. The issue is not the contribution guideline, which you do have also in the `CONTRIBUTING.rst` file. I meant something like https://www.astropy.org/code_of_conduct.html, more about the quality of interaction in the community than how to technically do things.
Some templates can also be found at https://help.github.com/en/articles/adding-a-code-of-conduct-to-your-project
