### 2019-10-02 79df574
Create CODE_OF_CONDUCT.rst
