### 2019-10-02 a456185
Deleted commented out code and unused function.
