### 2019-10-02 5e0ea02
Added tests to increase coverage.
