### 2019-10-02 aede36e
Made referee recommended changes to paper and citations.
