+++ CODE_OF_CONDUCT.rst
@@ -72,10 +72,12 @@ members of the project's leadership.
 Attribution
 -----------
 
-This Code of Conduct is adapted from the Contributor Covenant_, version 1.4,
-available at https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
+This Code of Conduct is adapted from the `Contributor Covenant`_, version 1.4,
+available at:
 
-:: _Contributor Covenant: https://www.contributor-covenant.org
+https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
+
+:: _`Contributor Covenant`: https://www.contributor-covenant.org
 
 For answers to common questions about this code of conduct, see
 https://www.contributor-covenant.org/faq
