+++ CODE_OF_CONDUCT.rst
@@ -77,7 +77,7 @@ available at:
 
 https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
 
-:: _`Contributor Covenant`: https://www.contributor-covenant.org
-
 For answers to common questions about this code of conduct, see
 https://www.contributor-covenant.org/faq
+
+:: _`Contributor Covenant`: https://www.contributor-covenant.org
