+++ CODE_OF_CONDUCT.rst
@@ -73,11 +73,12 @@ Attribution
 -----------
 
 This Code of Conduct is adapted from the `Contributor Covenant`_, version 1.4,
-available at:
+available here_.
 
-https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
+
 
 For answers to common questions about this code of conduct, see
 https://www.contributor-covenant.org/faq
 
-:: _`Contributor Covenant`: https://www.contributor-covenant.org
+.. _here: https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
+.. _`Contributor Covenant`: https://www.contributor-covenant.org
