### 2019-10-02 57e7116
Added code of conduct to docs.
