### 2019-10-02 #1775 @Hazboun6
I see. Thanks for the explanation and links @matteobachetti . That was very easy and I'm excited to include a Code of Conduct in the documents. I also went through and added a number of test for the code. I'm up to 87% coverage now. I've also added a few references and changed the `astropy` ref.
