### 2019-10-07 #1775 @cmbiwer
A Python package for building sensitivity curves for PTAs is presented. Overall, I believe the submission is suitable for publication in JOSS with minor comments to address. Going through the documentation, I found it to be readable and the functionality promised is presented. I do not believe any of these should prevent publication but that addressing them would strengthen the submission.

I outline my notes below for the authors:
  * While trying the tutorials, on the "Getting Started" page, you use ``numpy`` in the code blocks but omit it from the imports at the top of the page. I would suggest you add ``import numpy as np`` at the top of this page with the other imports. You do it for the following tutorials but not this page.
  * There was a minor typo in the first tutorial on ``GWBSensitivity`` and ``DeterSensitivity``. You wrote "inclusing" which I believe it was intended to be "including."
  * While running ``python setup.py test`` I received a number of warnings due to the LaTeX in the docstrings, e.g. ``DeprecationWarning: invalid escape sequence``. Although, I did not test myself (see documentation comment below), maybe these could be avoided by making the docstring ``r'''``. However, the tests ran successfully which is the important part.
  * In the tutorial on real data analysis (https://hasasia.readthedocs.io/en/latest/real_data_tutorial.html) a link to Lam et al. is included but does not seem to work.
  * In the documentation of two classes, there is an attribute with a docstring ``Hubble constant must be in`` which seems incomplete.
  * There was no documentation on how to build the documentation (any packages beyond the standard Sphinx?) and the documentation asks users to write documentation (https://hasasia.readthedocs.io/en/latest/contributing.html#write-documentation). Therefore, perhaps instructions on building the documentation could be added to this page.
  * In the software paper it states "In fact a number of 'standard' PTA configurations are included as part of the package." Is this referring to the NanoGrav 11-year data? Which seems to be committed to the repo and is also part of the tutorial? If so, perhaps this should be explicitly mentioned or cited in the software paper.
  * The accompanying theory paper by the authors have a nice set of citations in the introduction. Maybe there is an appropriate citation to review PTA here?
