+++ docs/real_data_tutorial.rst
@@ -103,7 +103,7 @@ Here we collate the noise parameters into one large dictionary.
 .. code:: python
 
     noise = {}
-    
+
     for nf in noise_files:
         with open(nf,'r') as fin:
             noise.update(json.load(fin))
@@ -150,7 +150,7 @@ Constructing the Correlation Matrix
 The following function makes a correlation matrix using the NANOGrav
 noise model and the parameters furnished in the data analysis release.
 For a detailed treatment of the noise modeling see `Lam, et al.,
-2015 <>`__.
+2015 <https://iopscience.iop.org/article/10.3847/1538-4357/834/1/35>`__.
 
 .. code:: python
 
@@ -171,7 +171,7 @@ For a detailed treatment of the noise modeling see `Lam, et al.,
             key_eq = '{0}_{1}_log10_{2}'.format(psr.name,be,'equad')
             sigma_sqr[mask] = (noise[key_ef]**2 * (psr.toaerrs[mask]**2)
                                + (10**noise[key_eq])**2)
-        
+
         corr = np.diag(sigma_sqr) + np.diag(u*ecorrs**2)
         return corr
 
@@ -233,12 +233,12 @@ needed to characterize frequencies much higher than investigated here.
         if ePsr.name in rn_psrs.keys():
             Amp, gam = rn_psrs[ePsr.name]
             plaw += hsen.red_noise_powerlaw(A=Amp, gamma=gam, freqs=freqs)
-            
+
         corr += hsen.corr_from_psd(freqs=freqs, psd=plaw,
                                    toas=ePsr.toas[::thin])
         psr = hsen.Pulsar(toas=ePsr.toas[::thin],
                           toaerrs=ePsr.toaerrs[::thin],
-                          phi=ePsr.phi,theta=ePsr.theta, 
+                          phi=ePsr.phi,theta=ePsr.theta,
                           N=corr, designmatrix=ePsr.Mmat[::thin,:])
         psr.name = ePsr.name
         psrs.append(psr)
@@ -320,7 +320,7 @@ highlighted.
             plt.loglog(sp.freqs,sp.h_c,lw=2,label=p.name)
         else:
             plt.loglog(sp.freqs,sp.h_c, color='k',lw=0.2)
-    
+
     plt.legend()
     plt.show()
     plt.close()
@@ -394,12 +394,12 @@ or index, or an array of them.
                                  SNR=SNR,
                                  S_eff=ng11yr_sc.S_eff)
     plaw_h = hgw*(ng11yr_sc.freqs/fyr)**(-2/3)
-    
-    #And for an array of alpha values. 
+
+    #And for an array of alpha values.
     alpha = np.linspace(-7/4,5/4,30)
     h=hsen.Agwb_from_Seff_plaw(freqs=ng11yr_sc.freqs,Tspan=Tspan,SNR=SNR,
                                S_eff=ng11yr_sc.S_eff,alpha=alpha)
-    
+
     plaw = np.dot((ng11yr_sc.freqs[:,np.newaxis]/fyr)**alpha,
                   h[:,np.newaxis]*np.eye(30))
 
@@ -409,7 +409,7 @@ or index, or an array of them.
         plt.loglog(ng11yr_sc.freqs,plaw[:,ii],
                    color='gray',lw=0.5)
     plt.loglog(ng11yr_sc.freqs,plaw_h,color='C1',lw=2,
-               label='SNR={0}, '.format(SNR)+r'$\alpha=-2/3$')  
+               label='SNR={0}, '.format(SNR)+r'$\alpha=-2/3$')
[…]
