### 2019-10-07 0e3f479
added r for doctstrings with LaTex.
