### 2019-10-07 af42fbc
added r for doctstrings with LaTex.
