### 2019-10-07 #1775 @Hazboun6
Thanks @cmbiwer for the detailed review! That `Deprecation error` appeared in the Python 3.6 and 3.7 builds on Travis, but I never saw them locally. The `r"""` did in fact fix things. I've fixed most of the points above. Regarding the last two points point, the newer version of the [manuscript](https://github.com/Hazboun6/hasasia/blob/master/paper/paper.md) mention the pickled NANOGrav sensitivity curve and includes two references to PTA reviews.
