### 2019-10-18 618b386
Adding Deterministic SNR for inc and sky averaged case
