### 2019-10-18 4488875
Adding in SNR for deterministic case and making some minor fixes/tests to it
