### 2019-10-22 557302b
Added JOSS Shield
