+++ HISTORY.rst
@@ -1,8 +1,10 @@
 =======
 History
 =======
+1.0.1 (2019-10-21)
+JOSS/Zenodo Release with various changes from JOSS Refereeing
 1.0.0 (2019-09-20)
-The Official Release. 
+The Official Release.
 0.1.6 (2019-08-30)
 0.1.6 (2019-08-29)
 0.1.5 (2019-08-13)
