### 2019-10-22 #1775 @Hazboun6
@arfon I'm just getting to this now. When you refer to the authors of the archive, do you mean the authors listed, for instance in an Authors.rst? Or am I missing something in Zenodo where one would list the authors?
