### 2019-10-22 #1775 @arfon
@Hazboun6 - Zenodo will attempt to fill out the author list based on the Git commit history. This is often not accurate so we ask that you make sure the Zenodo archive authors are the same as the authors on the paper.
