### 2019-10-22 ffea7d5
Changed version number and added zenodo json.
