### 2019-10-23 #1775 @Hazboun6
Alright, I think we are all set. Here is the badge for Zenodo [![DOI](https://zenodo.org/badge/178294587.svg)](https://zenodo.org/badge/latestdoi/178294587). I added a `.zenodo.json` so that the metadata file would reflect the authors of this paper. Please let me know what else you'll need me to do. Thanks!
