+++ .travis.yml
@@ -1,26 +1,4 @@
 # This file was autogenerated and will overwrite each time you run travis_pypi_setup.py
-deploy:
-  provider: pypi
-  distributions: sdist bdist_wheel
-  user: Hazboun6
-  password:
-    secure: !!binary |
-      VldGdDJpRkFaT3FnZjh2aVM4dnEzb1lVN3lYbVArc01IdDhvTmRhYkZGWVpMYk0vSlBtaDk5eGFM
-      S0RnZStoY1BkdG9HUndyT0dkdm53RjUzckNiZXBWQXRPbDRxNldVSjZQR2svaVYzSmVLQU1iSlpv
-      d3NQYW9KOHdoclJjbVdBeGExOUE1TEwwcnJNdnBRZUJuSE93YS9RY2pNcmJ1dmh1djJkcHF3VWM0
-      dVN4cEtzKzdsVjhsYnFJV0hTeDlUSjV1S2Fyc1pkamRGL3Y3T2ltV2l2UDlyYUtxR1BZeURlV21O
-      NEVWMEdyT0gxZmxGV0RGVXZZUGc0N3FWWDVDdlFUellIcjUzVmd4U3VBcngzRUd6YXE0VFlEMFpq
-      ZlorcHdtUHZuUWMrR1ZRaTFFNFhiaDVyMXF2ajdkVS9RY0hTVkxoaTgxQ2Mzem9ET2VJRFo3c1Vp
-      L1g1d0FHSldPZTdHUUpwOU1zU1lsVEdQTHB4ZVdCNUlyeU9iMEluaHkwcTNtZjkwcUFzK040UC9Q
-      NFU2VWxiS05jeHdIMVc2YmJGanhuaExKQXlmR3QzVVZVNllkNzVZVXVDN1hZUUpGaFp6OUQzOFVM
-      bkdlYUdHakJMVVNSQ3ozTUFpSy9UMms1RndVTDBWSDVtbnZJZUFVZDkvd095bGh5N3FadDFEZUVi
-      aTR2UFJOMC9qNjFxcURLYVY0cndXYkZkd3lkUjFZeEZIS0ErQk5mb1dQRzE2elF2dFlGVDhiN1ht
-      bWxycmJqeXdOdkVCYWhNYmI4WGErdjdrV1JEeDZEUUVOYzVWdFA0UUwzNXRUcE1VUEkwN3N1TU14
-      VXhkYVh4aFZvRHZNbWttT2ZScTV0SE5Bd0p5Q0VvS2FXRS9tY29lay9DM2hWckRnNlRnQTV3Z2M9
-  true:
-    python: 3.7
-    repo: Hazboun6/hasasia
-    tags: true
 install: pip install -U tox-travis
 language: python
 python:
