### 2019-10-23 #1775 @arfon
Looks like the DOI isn't resolving yet. I'll hold off closing this thread until that's working properly...
